+++ foyer-storage/src/large/generic.rs
@@ -183,6 +183,14 @@ where
     S: HashBuilder + Debug,
 {
     async fn open(mut config: GenericLargeStorageConfig<K, V, S>) -> Result<Self> {
+        if config.flushers + config.clean_region_threshold > config.device.regions() / 2 {
+            tracing::warn!("[lodc]: large object disk cache stable regions count is too small, flusher [{flushers}] + clean region threshold [{clean_region_threshold}] (default = reclaimers) is supposed to be much larger than the region count [{regions}]",
+                flushers = config.flushers,
+                clean_region_threshold = config.clean_region_threshold,
+                regions = config.device.regions()
+            );
+        }
+
         let stats = config.statistics.clone();
 
         let device = config.device.clone();
@@ -525,6 +533,7 @@ mod tests {
 
     fn cache_for_test() -> Cache<u64, Vec<u8>> {
         CacheBuilder::new(10)
+            .with_shards(1)
             .with_eviction_config(FifoConfig::default())
             .build()
     }