+++ foyer/src/prelude.rs
@@ -16,7 +16,7 @@ pub use crate::{
     common::{
         buf::{BufExt, BufMutExt},
         code::{Key, StorageKey, StorageValue, Value},
-        event::EventListener,
+        event::{Event, EventListener},
         range::RangeBoundsExt,
         tracing::TracingOptions,
     },
@@ -26,7 +26,7 @@ pub use crate::{
         writer::{HybridCacheStorageWriter, HybridCacheWriter},
     },
     memory::{
-        Cache, CacheBuilder, CacheContext, CacheEntry, EvictionConfig, FetchState, FifoConfig, LfuConfig, LruConfig,
+        Cache, CacheBuilder, CacheEntry, CacheHint, EvictionConfig, FetchState, FifoConfig, LfuConfig, LruConfig,
         S3FifoConfig, Weighter,
     },
     storage::{